added in some CLI args

+++ prep_data.py
@@ -2,34 +2,51 @@ import zipfile
 import torchvision
 import os
 from PIL import Image
+import numpy as np
+import argparse
 
-### Unpack and Save CatsVsDogs ###
-print("Unpacking CatsvsDogs")
-with zipfile.ZipFile("data/kagglecatsanddogs_5340.zip", "r") as zip:
-    zip.extractall("data")
-    
-### Clean Up CatsVDogs ###
-path_to_cats = os.path.join("../data/PetImages/", "Cat") # Get Path to Cat folder
-path_to_dogs = os.path.join("../data/PetImages/", "Dog") # Get Path to Dog folder 
+parser = argparse.ArgumentParser("Data Preperation Script")
 
-dog_files = os.listdir(path_to_dogs) # Get list of all files inside of dog folder
-cat_files = os.listdir(path_to_cats) # Get list of all files inside cat folder
+parser.add_argument("-p", "--path", type=str, required=True, help="Path to Data Root")
+parser.add_argument("--all", action="store_true")
+parser.add_argument("--catsvdogs", action="store_true")
+parser.add_argument("--mnist", action="store_true")
 
-path_to_dog_files = [os.path.join(path_to_dogs, file) for file in dog_files] # Get full path to each cat file
-path_to_cat_files = [os.path.join(path_to_cats, file) for file in cat_files] # Get full path to each dog file
+args = parser.parse_args()
 
-path_to_files = path_to_dog_files + path_to_cat_files
+if args.all is True:
+    args.catsvdogs = True
+    args.mnist = True
 
-for file in path_to_files:
-    try:    
-        img = np.array(Image.open(file))
-        if img.shape[-1] != 3: 
-            os.remove(file) # Delete image if it doesnt have three channels
-    except:
-        os.remove(file) # Delete if not an image file, or broken image
+if args.catsvdogs:
+    ### Unpack and Save CatsVsDogs ###
+    print("Unpacking CatsvsDogs")
+    with zipfile.ZipFile(f"{args.path}/kagglecatsanddogs_5340.zip", "r") as zip:
+        zip.extractall(args.path)
 
-### Download MNST Dataset ###
-train = torchvision.datasets.MNIST('data', train=True, download=True)
-test = torchvision.datasets.MNIST('data', train=False, download=True)
+    ### Clean Up CatsVDogs ###
+    path_to_cats = os.path.join(f"{args.path}/PetImages/", "Cat") # Get Path to Cat folder
+    path_to_dogs = os.path.join(f"{args.path}/PetImages/", "Dog") # Get Path to Dog folder
+
+    dog_files = os.listdir(path_to_dogs) # Get list of all files inside of dog folder
+    cat_files = os.listdir(path_to_cats) # Get list of all files inside cat folder
+
+    path_to_dog_files = [os.path.join(path_to_dogs, file) for file in dog_files] # Get full path to each cat file
+    path_to_cat_files = [os.path.join(path_to_cats, file) for file in cat_files] # Get full path to each dog file
+
+    path_to_files = path_to_dog_files + path_to_cat_files
+
+    for file in path_to_files:
+        try:
+            img = np.array(Image.open(file))
+            if img.shape[-1] != 3:
+                os.remove(file) # Delete image if it doesnt have three channels
+        except:
+            os.remove(file) # Delete if not an image file, or broken image
+
+if args.mnist:
+    ### Download MNST Dataset ###
+    train = torchvision.datasets.MNIST(args.path, train=True, download=True)
+    test = torchvision.datasets.MNIST(args.path, train=False, download=True)
